unit_2_final/requirements.txt
confluent-kafka==2.11.0
faust-streaming==0.11.3